Remove Form 230 fill instruction video :speech_balloon:, #54
The current size of the video (~ 14Mb), even compressed, does not justify
the values it adds. Even dropping irrelevant frames and making a gif
only reduces the size to ~ 3Mb which still does not seem to be
justified. Other alternatives to reduce the size like editing the video
further and keeping it in video format to compress better were considered
but the time investment is not justified.
Video on how to fill forms using Adobe Reader or other PDF editor can be
easily found on the internet.